Bugfix
Fixed an issue in race manager : RaceState is 4 when the race is finished, not 3. (atleast for TTs). It was previously causing the info display to turn off on the race finish state, despite the code not aiming to do that.

Fixed the Rotation speed calculation to not trigger error while not in a race. (It wasn't a big issue, it was just spamming the python log)